Shader and VFX Graph integration
This uses a depreciated feature of the shader graph (visual effect targeting) but I can't find a newer replacement. Will stay up to day on developments in Unity to find a suitable replacement workflow for shader and vfx integration.

Feature:
Wrote a shader graph which has a scrolling alpha texture for a laser beam, this shader is then used in the particle strip of the vfx graph for the grapple hook laser to produce a animated energy beam look.

BUG: HDR colors are broken, appear to change intensity whenever the color picker is opened and closed, making it difficult to get the bloom/glow effect right.